+++ public/atompagebuilder/app/_app/immutable/chunks/1-14eb57f5.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/error.svelte-f8ec66c3.js";export{t as component};