+++ public/atompagebuilder/app/_app/immutable/chunks/2-6384e557.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_layout.svelte-1ddc74da.js";export{t as component};