+++ public/atompagebuilder/app/_app/immutable/chunks/3-b0b604ee.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(editor)/_layout.svelte-71339d48.js";export{t as component};