+++ public/atompagebuilder/app/_app/immutable/chunks/4-eb5be93c.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_page.svelte-ef5c99fe.js";export{t as component};