+++ public/atompagebuilder/app/_app/immutable/chunks/5-4d1c139c.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(editor)/editor/_page.svelte-6a00022a.js";export{t as component};